fixed some classes according to comments
