+++ pom.xml
@@ -57,5 +57,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
